[Refactor] Decouple PipelineDriver observer handling
