+++ be/src/exec/CMakeLists.txt
@@ -178,7 +178,6 @@ set(PIPELINE_RUNTIME_SRCS
     pipeline/query_context_manager.cpp
     pipeline/schedule/timeout_tasks.cpp
     pipeline/schedule/event_scheduler.cpp
-    pipeline/schedule/pipeline_driver_observer.cpp
     pipeline/runtime_filter_types.cpp
 )
